testing aggregate collection using query rpc call

+++ services/core/SQLAggregateHistorian/tests/test_aggregator.py
@@ -19,7 +19,7 @@ except:
     HAS_MYSQL_CONNECTOR = False
 
 # table_defs with prefix
-sqlite_platform = {
+sqlite_aggregator = {
     "agentid": "aggregate-sqlite",
     "identity": "aggregate-hist-sqlite",
     "connection": {
@@ -72,7 +72,7 @@ sqlite_platform = {
 # "historian" and grant historian user access to "historian" database
 
 # table_defs with prefix
-mysql_platform = {
+mysql_aggregator = {
     "agentid": "aggregate-mysql",
     "identity": "aggregate-hist-mysql",
     "connection": {
@@ -137,16 +137,14 @@ meta_table = 'meta'
 
 
 @pytest.fixture(scope="module")
-def publish_agent(request, volttron_instance):
-    # 1: Start a fake agent to publish to message bus
-    print("**In setup of publish_agent volttron is_running {}".format(
-        volttron_instance.is_running))
+def query_agent(request, volttron_instance):
+    # 1: Start a fake agent to query the sqlhistorian in volttron_instance2
     agent = volttron_instance.build_agent()
 
     # 2: add a tear down method to stop sqlhistorian agent and the fake
     # agent that published to message bus
     def stop_agent():
-        print("In teardown method of publish_agent")
+        print("In teardown method of query_agent")
         agent.core.stop()
 
     request.addfinalizer(stop_agent)
@@ -158,10 +156,10 @@ def publish_agent(request, volttron_instance):
                 params=[
                     pytest.mark.skipif(
                         not HAS_MYSQL_CONNECTOR,
-                        reason='No mysql client available.')(mysql_platform),
-                    sqlite_platform
+                        reason='No mysql client available.')(mysql_aggregator),
+                    sqlite_aggregator
                 ])
-def aggregate_agent(request, volttron_instance1):
+def aggregate_agent(request, volttron_instance):
     global db_connection, data_table, \
         topics_table, meta_table
     print("** Setting up test_sqlhistorian module **")
@@ -170,35 +168,28 @@ def aggregate_agent(request, volttron_instance1):
     print("request param", request.param)
     if request.param['connection']['type'] == 'sqlite':
         request.param['connection']['params']['database'] = \
-            volttron_instance1.volttron_home + "/historian.sqlite"
-
-    # Make database connection
-    agent_uuid = volttron_instance1.install_agent(
-        agent_dir="services/core/SQLAggregateHistorian",
-        config_file=request.param,
-        start=False)
-    print("agent id: ", agent_uuid)
+            volttron_instance.volttron_home + "/historian.sqlite"
 
     # figure out db table names from config
-    # Set this hear so that cleanup fixture can use it
+    # Set this hear so that we can create these table after connecting to db
     if request.param.get('tables_def', None) is None:
         data_table = 'data'
         topics_table = 'topics'
         meta_table = 'meta'
     elif request.param['tables_def']['table_prefix']:
         data_table = request.param['tables_def']['table_prefix'] + "_" + \
-            request.param['tables_def']['data_table']
+                     request.param['tables_def']['data_table']
         topics_table = request.param['tables_def']['table_prefix'] + "_" + \
-            request.param['tables_def']['topics_table']
+                       request.param['tables_def']['topics_table']
         meta_table = request.param['tables_def']['table_prefix'] + "_" + \
-            request.param['tables_def']['meta_table']
+                     request.param['tables_def']['meta_table']
     else:
         data_table = request.param['tables_def']['data_table']
         topics_table = request.param['tables_def']['topics_table']
         meta_table = request.param['tables_def']['meta_table']
 
     # 2: Open db connection that can be used for row deletes after
-    # each test method. Create tables in case of mysql
+    # each test method. Create tables and insert test topics in topics table
     if request.param['connection']['type'] == "sqlite":
         connect_sqlite(request)
     elif request.param['connection']['type'] == "mysql":
@@ -209,6 +200,25 @@ def aggregate_agent(request, volttron_instance1):
         pytest.fail(msg="Invalid database type specified " +
                         request.param['connection']['type'])
 
+    # 2. Install agents - sqlhistorian, sqlaggregatehistorian
+    config = dict()
+    config['agentid'] = 'sqlhistorian'
+    config['identity'] = 'platform.historian'
+    config['connection'] = request.param['connection']
+    config['tables_def'] = request.param['tables_def']
+
+    sql_historian_uuid = volttron_instance.install_agent(
+        agent_dir="services/core/SQLHistorian",
+        config_file=config,
+        start=True)
+    print("agent id: ", sql_historian_uuid)
+
+    agent_uuid = volttron_instance.install_agent(
+        agent_dir="services/core/SQLAggregateHistorian",
+        config_file=request.param,
+        start=False)
+    print("agent id: ", agent_uuid)
+
     # 3: add a tear down method to stop sqlhistorian agent and the fake
     # agent that published to message bus
     def stop_agent():
@@ -216,8 +226,9 @@ def aggregate_agent(request, volttron_instance1):
         if db_connection:
             db_connection.close()
             print("closed connection to db")
-        if volttron_instance1.is_running():
-            volttron_instance1.remove_agent(agent_uuid)
+        if volttron_instance.is_running():
+            volttron_instance.remove_agent(agent_uuid)
+            volttron_instance.remove_agent(sql_historian_uuid)
 
     request.addfinalizer(stop_agent)
     return agent_uuid
@@ -341,7 +352,7 @@ def publish_test_data(start_time, start_reading, count):
 
 
 @pytest.mark.aggregator
-def test_basic_function(aggregate_agent, volttron_instance1):
+def test_basic_function(volttron_instance, aggregate_agent, query_agent):
     """
 
     @param aggregate_agent:
@@ -355,32 +366,54 @@ def test_basic_function(aggregate_agent, volttron_instance1):
     publish_test_data(start_time, 0, 5)
     gevent.sleep(0.5)
     cursor = db_connection.cursor()
-    if isinstance(db_connection, mysql.connection.MySQLConnection):
+    if HAS_MYSQL_CONNECTOR and isinstance(db_connection, \
+                                          mysql.connection.MySQLConnection):
         cursor.execute("DELETE FROM sum_2m")
+        cursor.execute("DELETE FROM sum_3m")
         db_connection.commit()
-    volttron_instance1.start_agent(aggregate_agent)
+    volttron_instance.start_agent(aggregate_agent)
     # gevent.sleep(2)
     gevent.sleep(5 * 60)  # sleep till we see two rows in aggregate table
-    cursor = db_connection.cursor()
-    cursor.execute("SELECT value_string from sum_2m "
-                   "WHERE topic_id =1")
-    rows = cursor.fetchall()
-    print ("result is {}".format(rows))
-    assert float(rows[0][0]) == 3.0
-    assert float(rows[1][0]) == 7.0
 
-    cursor = db_connection.cursor()
-    cursor.execute("SELECT value_string from sum_2m where "
-                   "topic_id =2")
-    rows = cursor.fetchall()
-    print ("result is {}".format(rows))
-    assert float(rows[0][0]) == 3.0
-    assert float(rows[1][0]) == 7.0
+    result = query_agent.vip.rpc.call('platform.historian',
+                                      'query',
+                                      topic='device1/out_temp',
+                                      agg_type='sum',
+                                      agg_period='2m',
+                                      count=20,
+                                      order="FIRST_TO_LAST").get(timeout=100)
 
-    cursor.execute("SELECT value_string from sum_3m where "
-                   "topic_id =1")
-    rows = cursor.fetchall()
-    print ("result is {}".format(rows))
-    assert float(rows[0][0]) == 3.0
-    assert float(rows[1][0]) == 7.0
+    print(result)
+    assert (result['values'][0][1] == 3.0)
+    assert (result['values'][1][1] == 7.0)
 
+    result = query_agent.vip.rpc.call('platform.historian',
+                                      'query',
+                                      topic='device1/in_temp',
+                                      agg_type='sum',
+                                      agg_period='2m',
+                                      count=20,
+                                      order="FIRST_TO_LAST").get(timeout=100)
+    assert (result['values'][0][1] == 3.0)
+    assert (result['values'][1][1] == 7.0)
+
+    result = query_agent.vip.rpc.call('platform.historian',
+                                      'query',
+                                      topic='device1/in_temp',
+                                      agg_type='sum',
+                                      agg_period='3m',
+                                      count=20,
+                                      order="FIRST_TO_LAST").get(timeout=100)
+    assert (result['values'][0][1] == 3.0)
+    assert (result['values'][1][1] == 7.0)
+
+    result = query_agent.vip.rpc.call('platform.historian',
+                                      'query',
+                                      topic='device1/out_temp',
+                                      agg_type='sum',
+                                      agg_period='3m',
+                                      count=20,
+                                      order="FIRST_TO_LAST").get(timeout=100)
+
+    assert (result['values'][0][1] == 3.0)
+    assert (result['values'][1][1] == 7.0)
